pyproject.toml
[project]
name = "homepage"
version = "0.1.0"
description = "Personal Website"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi",
    "boto3",
    "cast-vue",
    "cast-bootstrap5",
    "django-resume",
    "django-allauth",
    "django-anymail[mailgun]",
    "django-extensions",
    "django-indieweb",
    "django-proxy",
    "django-storages",
    "django-cast",
    "django-vite",
    "gunicorn",
    "httpx",
    "psycopg[binary]",
    "python-dotenv",
    "pyzmq>=25.1.1",
    "sentry-sdk",
    "whitenoise",
]

[tool.uv]
dev-dependencies = [
    "black",
    "coverage",
    "factory-boy",
    "flake8",
    "flit",
    "django-debug-toolbar",
    "django-test-plus",
    "dj-notebook",
    "pytest",
    "pytest-sugar",
    "pytest-django",
    "rich",
    "typer",
    "django-removals",
]

[tool.uv.sources]
cast-vue = { git = "https://github.com/ephes/cast-vue" }
cast-bootstrap5 = { git = "https://github.com/ephes/cast-bootstrap5" }
django-cast = { git = "https://github.com/ephes/django-cast", branch="develop" }
django-indieweb = { path = "../django-indieweb", editable = true }

[tool.black]
line-length = 119

[tool.isort]
profile = "black"

[tool.coverage.run]
branch = true
source = ["apps"]
omit = ["apps/*/tests/*", "apps/*/migrations/*"]
command_line = "-m pytest"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
filterwarnings = [
    # "ignore::django.utils.deprecation.RemovedInDjango60Warning",
]
addopts = [
    "-ra",
    "--reuse-db",
    "--no-migrations",
]
